#162: use deleteReply to consolidate viewgearcollector flow
